fix: added flag for run.sh to not always recreate db

+++ run.sh
@@ -8,8 +8,13 @@ echo "Environment configured!"
 
 echo "Starting up docker container!"
 
-# Force run init.sql each run
-docker rm "$DATABASE_CONTAINER" 2>/dev/null || true
+if [ "$1" = "-f" ]; then
+    # Force run init.sql each run
+    echo "Forcing database container recreation"
+    docker rm "$DATABASE_CONTAINER" 2>/dev/null || true
+else
+  echo "Not forcing database container recreation"
+fi
 docker-compose up --build --force-recreate --always-recreate-deps -d
 cargo build --release
